% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_private.R
\name{a_add_suffix}
\alias{a_add_suffix}
\title{Add suffix to corresponding biomarker}
\usage{
a_add_suffix(dat, suff)
}
\arguments{
\item{dat}{A data frame that contains all the information for 
variables of specific biomarkers.}

\item{suff}{A string indicating the type of biomarker (i.e. _cog).}
}
\value{
A data frame that contains all the information for 
    variables of specific biomarkers with suffix appended to each
    variable.
}
\description{
Add suffix to corresponding biomarker
}
