% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_private.R
\name{a_cut_variable}
\alias{a_cut_variable}
\title{Cutting continuous variables to categorical variables}
\usage{
a_cut_variable(dat, var, n)
}
\arguments{
\item{dat}{The analysis dataset.}

\item{var}{The continuous variables in analysis dataset.}

\item{n}{The number of categories.}
}
\value{

}
\description{
Cutting continuous variables to categorical variables
}
