% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_private.R
\name{a_select_bmk}
\alias{a_select_bmk}
\title{Select biomarkers with corresponding window restriction}
\usage{
a_select_bmk(visit, dat, win, window_overlap)
}
\arguments{
\item{visit}{The reference time with patients' ID}

\item{dat}{A data frame that contains all the information for 
variables of specific biomarkers.}

\item{win}{A integer (unit of days) indicating the maximum 
acceptable gap time for merging a biomarker test to base data.}

\item{window_overlap}{A logical value indicating whether the 
time window should overlap.}
}
\value{
A data frame with biomarkers for each patient in every reference time.
}
\description{
Select biomarkers with corresponding window restriction
}
