% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_vector_to_matrix_column}
\alias{add_vector_to_matrix_column}
\title{Add a column vector to each column of a matrix}
\usage{
add_vector_to_matrix_column(v0, m0)
}
\arguments{
\item{v0}{Column vector}

\item{m0}{Matrix}
}
\description{
Add a column vector to each column of a matrix
}
