% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_vector_to_matrix_row}
\alias{add_vector_to_matrix_row}
\title{Add a column vector to each row of a matrix}
\usage{
add_vector_to_matrix_row(v0, m0)
}
\arguments{
\item{v0}{Column vector}

\item{m0}{Matrix}
}
\description{
Add a column vector to each row of a matrix
}
