% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_arithmetic_1_cross_class.R
\name{-,ANY,dual-method}
\alias{-,ANY,dual-method}
\title{Subtraction of 'dual'-class objects (dual-ANY)}
\usage{
\S4method{-}{ANY,dual}(e1, e2)
}
\arguments{
\item{e1}{"ANY" object.}

\item{e2}{A "dual" object.}
}
\description{
Subtraction of 'dual'-class objects (dual-ANY)
}
