% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_generic.R
\name{as.matrix.dual}
\alias{as.matrix.dual}
\title{Coerce the first component of the dual object into a matrix.}
\usage{
\method{as.matrix}{dual}(x, ...)
}
\arguments{
\item{x}{A "dual" object.}

\item{...}{Use `nrow`, `ncol` to specify the number of rows and columns.}
}
\description{
Coerce the first component of the dual object into a matrix.
}
