% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_structure_reshape.R
\name{as.vector,dual-method}
\alias{as.vector,dual-method}
\title{Coerce the first component of the dual object into a vector.}
\usage{
\S4method{as.vector}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Coerce the first component of the dual object into a vector.
}
