% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_diff.R
\name{auto_diff}
\alias{auto_diff}
\title{Automatic differentiation}
\usage{
auto_diff(f, vary, fix = NULL)
}
\arguments{
\item{f}{A function of which the derivative is seeked.}

\item{vary}{A named list of variables; the variables to be varied.}

\item{fix}{A named list of variables; the variables to be fixed.}
}
\description{
Automatic differentiation
}
\examples{
f <- function(y, X, beta) { y - X \%*\% beta }
auto_diff(f,
  vary = list(beta = c(5,6)),
  fix = list(X = matrix(1:4, 2, 2), y = c(2,3))
)

g <- function(X, Y) { X \%*\% Y }
X <- randn(2, 2)
Y <- randn(2, 2)
auto_diff(g, vary = list(X = X, Y = Y))

}
