% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_special_matrix.R
\name{band_matrix}
\alias{band_matrix}
\title{Band matrix}
\usage{
band_matrix(nr, nc, x = 1)
}
\arguments{
\item{nr}{integer; row dimension.}

\item{nc}{integer; column dimension.}

\item{x}{A scalar or a vector to be placed on the diagonal of the matrix.}
}
\description{
Band matrix
}
\examples{
band_matrix(2, 3)
band_matrix(2, 3, c(999))
band_matrix(2, 3, c(999, 111))

}
