% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_kronecker_product.R
\name{BxAZ}
\alias{BxAZ}
\title{Post-multiplying a kronecker product}
\usage{
BxAZ(B, A, Z)
}
\arguments{
\item{B}{numeric matrix.}

\item{A}{numeric matrix.}

\item{Z}{numeric matrix.}
}
\description{
Compute (B \%x\% A) Z
}
