% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_binding.R
\docType{methods}
\name{cbind2,dual,dual-method}
\alias{cbind2,dual,dual-method}
\title{Combine 'dual'-class objects by Columns}
\usage{
\S4method{cbind2}{dual,dual}(x, y)
}
\arguments{
\item{x}{A "dual" object.}

\item{y}{A "dual" object.}
}
\description{
Combine 'dual'-class objects by Columns
}
