% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_matrix_list.R
\docType{methods}
\name{cbind2,matrix_list,matrix_list-method}
\alias{cbind2,matrix_list,matrix_list-method}
\title{Combine two 'matrix_list'-class objects by Columns}
\usage{
\S4method{cbind2}{matrix_list,matrix_list}(x, y)
}
\arguments{
\item{x}{A "matrix_list" object.}

\item{y}{A "matrix_list" object.}
}
\description{
Combine two 'matrix_list'-class objects by Columns
}
