% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_matrix_ops.R
\name{chol0,dual-method}
\alias{chol0,dual-method}
\title{Cholesky decomposition of 'dual'-class objects}
\usage{
\S4method{chol0}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Cholesky decomposition of 'dual'-class objects
}
\note{
The Cholesky decomposition used in this function returns a
lower triangular matrix.
}
