% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_math_sum.R
\name{colMeans,dual-method}
\alias{colMeans,dual-method}
\title{Column mean of a matrix.}
\usage{
\S4method{colMeans}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Column mean of a matrix.
}
