% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sum.R
\docType{methods}
\name{colSums,dual-method}
\alias{colSums,dual-method}
\title{Column sum of a matrix.}
\usage{
\S4method{colSums}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Column sum of a matrix.
}
