% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_special_matrix.R
\name{commutation_matrix}
\alias{commutation_matrix}
\title{Commutation matrix}
\usage{
commutation_matrix(r, c)
}
\arguments{
\item{r}{integer; row dimension.}

\item{c}{integer; column dimension.}
}
\description{
Commutation matrix
}
