% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{commutation_matrix0}
\alias{commutation_matrix0}
\title{Commutation matrix}
\usage{
commutation_matrix0(...)
}
\arguments{
\item{r}{integer; row dimension.}

\item{c}{integer; column dimension.}
}
\description{
Commutation matrix
}
