% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_math_elementary.R
\name{cos,dual-method}
\alias{cos,dual-method}
\title{Element-wise cosine of a dual object}
\usage{
\S4method{cos}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Element-wise cosine of a dual object
}
