% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_matrix_ops.R
\name{crossprod,dual,missing-method}
\alias{crossprod,dual,missing-method}
\alias{crossprod,dual,ANY-method}
\alias{crossprod,dual,dual-method}
\title{Crossproduct of 'dual'-class objects}
\usage{
\S4method{crossprod}{dual,missing}(x, y)

\S4method{crossprod}{dual,ANY}(x, y)

\S4method{crossprod}{dual,dual}(x, y)
}
\arguments{
\item{x}{A numeric matrix, or the corresponding "dual" object.}

\item{y}{A numeric matrix, or the corresponding "dual" object.}
}
\description{
Crossproduct of 'dual'-class objects
}
