% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_density.R
\name{dchisq0}
\alias{dchisq0}
\title{The density of the chi-squared distribution}
\usage{
dchisq0(x, df, log = FALSE)
}
\arguments{
\item{x}{vector or matrix of quantiles. If x is a matrix, each row is taken to be a quantile.}

\item{df}{degrees of freedom (> 0, maybe non-integer).}

\item{log}{logical; if TRUE, returns the log value.}
}
\description{
The density of the chi-squared distribution
}
\examples{
n <- 10
x <- rchisq0(n, df = 3)
dchisq0(x, df = 3)

}
