% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_matrix_ops.R
\name{det,dual-method}
\alias{det,dual-method}
\title{Determinant of a 'dual'-class object}
\usage{
\S4method{det}{dual}(x, ...)
}
\arguments{
\item{x}{A "dual" object.}

\item{...}{Other parameters to be passed to `base::det`.}
}
\description{
Determinant of a 'dual'-class object
}
