% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_matrix_ops.R
\name{det0}
\alias{det0}
\title{Determinant of a matrix
This is a wrapper of `base::det`.}
\usage{
det0(x, ...)
}
\arguments{
\item{x}{A numeric matrix.}

\item{...}{Other parameters to be passed to `base::det`.}
}
\description{
Determinant of a matrix
This is a wrapper of `base::det`.
}
