% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_density.R
\name{dexp0}
\alias{dexp0}
\title{The density of the exponential distribution}
\usage{
dexp0(x, rate = 1, log = FALSE)
}
\arguments{
\item{x}{vector or matrix of quantiles. If x is a matrix, each row is taken to be a quantile.}

\item{rate}{positive number; the rate parameter.}

\item{log}{logical; if TRUE, returns the log value.}
}
\description{
The density of the exponential distribution
}
\examples{
n <- 10
x <- rexp0(n, rate = 3)
dexp0(x, rate = 3)

}
