% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_density.R
\name{dgamma0}
\alias{dgamma0}
\title{The density of the gamma distribution}
\usage{
dgamma0(x, shape, rate, scale = 1/rate, log = FALSE)
}
\arguments{
\item{x}{vector or matrix of quantiles. If x is a matrix, each row is taken to be a quantile.}

\item{shape}{positive number; the shape parameter.}

\item{rate}{positive number; the rate parameter.}

\item{scale}{positive number; the scale parameter.}

\item{log}{logical; if TRUE, returns the log value.}
}
\description{
The density of the gamma distribution
}
\examples{
n <- 10
x <- rgamma0(n, shape = 1, scale = 5)
dgamma0(x, shape = 1, scale = 5)

}
