% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_structure_reshape.R
\name{diag}
\alias{diag}
\title{Diagonal matrix}
\usage{
diag(x = 1, nrow, ncol, names = TRUE)
}
\arguments{
\item{x}{A matrix, a vector or a dual number. In the last case, all other parameters
are ignored as \code{diag.dual} only takes the first argument.}

\item{nrow}{optional dimensions for the result when \code{x} is
    not a matrix.}

\item{ncol}{optional dimensions for the result when \code{x} is
    not a matrix.}

\item{names}{(when \code{x} is a matrix) logical indicating if the
    resulting vector, the diagonal of \code{x}, should inherit
    \code{\link{names}} from \code{dimnames(x)} if available.}
}
\description{
Diagonal matrix
}
