% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_structure_reshape.R
\docType{methods}
\name{diag.dual}
\alias{diag.dual}
\alias{diag,dual-method}
\title{Diagonal matrix}
\usage{
diag.dual(x)

\S4method{diag}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Diagonal matrix
}
