% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_special_matrix.R
\name{Diagonal}
\alias{Diagonal}
\title{Diagonal matrix}
\usage{
Diagonal(n, x = NULL)
}
\arguments{
\item{n}{integer; the dimension of the square matrix.}

\item{x}{A scalar or a vector to be placed on the diagonal of the matrix.}
}
\description{
Diagonal matrix
}
\examples{
Diagonal(3)
Diagonal(3, 999)
Diagonal(3, c(11, 22, 33))

}
