% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_special_matrix.R
\name{Diagonal0}
\alias{Diagonal0}
\title{Diagonal matrix}
\usage{
Diagonal0(...)
}
\arguments{
\item{...}{`n` integer; the dimension of the square matrix.

`x` A scalar or a vector to be placed on the diagonal of the matrix.}
}
\description{
Diagonal matrix
}
