% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_density.R
\name{dt0}
\alias{dt0}
\title{The density of the student-t distribution}
\usage{
dt0(x, df, log = FALSE)
}
\arguments{
\item{x}{vector or matrix of quantiles. If x is a matrix, each row is taken to be a quantile.}

\item{df}{degrees of freedom (> 0, maybe non-integer).}

\item{log}{logical; if TRUE, returns the log value.}
}
\description{
The density of the student-t distribution
}
\examples{
n <- 10
x <- rt0(n, df = 3)
dt0(x, df = 3)

}
