% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_arithmetic_1_cross_class.R
\name{-,dual,missing-method}
\alias{-,dual,missing-method}
\title{Subtraction of 'dual'-class objects (unary dual)}
\usage{
\S4method{-}{dual,missing}(e1, e2)
}
\arguments{
\item{e1}{A "dual" object.}

\item{e2}{"missing" object.}
}
\description{
Subtraction of 'dual'-class objects (unary dual)
}
