% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dual_def.R
\name{dual}
\alias{dual}
\title{Dual number constructor}
\usage{
dual(x, param_dim, ind)
}
\arguments{
\item{x}{The object to be converted to a dual number.}

\item{param_dim}{The dimension of the dual component to be attached.}

\item{ind}{Integer; the index of the parameter identical to x.}
}
\description{
Dual number constructor
}
