% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_diff.R
\name{dual_list}
\alias{dual_list}
\title{Converting a list of parameters into a list of dual numbers}
\usage{
dual_list(ls_params)
}
\arguments{
\item{ls_params}{A named list of parameters}
}
\description{
Converting a list of parameters into a list of dual numbers
}
\examples{
\dontrun{
X <- randn(2, 2)
y <- randn(2)
dual_list(list(X = X, y = y))
}
}
