% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_diff.R
\name{duals}
\alias{duals}
\title{Converting a list of parameters into a list of dual numbers}
\usage{
duals(vary)
}
\arguments{
\item{vary}{A named list of parameters}
}
\description{
Converting a list of parameters into a list of dual numbers
}
\examples{
\dontrun{
X <- randn(2, 2)
y <- rnorm(2)
duals(list(X = X, y = y))
}
}
