% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_density.R
\name{dWishart0}
\alias{dWishart0}
\title{The density of the Wishart distribution}
\usage{
dWishart0(X, df, Scale, log = FALSE)
}
\arguments{
\item{X}{numeric matrix.}

\item{df}{degrees of freedom (> 0, maybe non-integer).}

\item{Scale}{numeric matrix; the scale matrix.}

\item{log}{logical; if TRUE, returns the log value.}
}
\description{
The density of the Wishart distribution
}
