% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_special_matrix.R
\name{elimination_matrix}
\alias{elimination_matrix}
\title{Elimination matrix}
\usage{
elimination_matrix(n)
}
\arguments{
\item{n}{integer, row or column dimension.}
}
\description{
Elimination matrix
}
