% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_special_matrix.R
\name{elimination_matrix0}
\alias{elimination_matrix0}
\title{Elimination matrix (memoised)}
\usage{
elimination_matrix0(...)
}
\arguments{
\item{...}{`n` integer, row or column dimension.

Use empty argument to clear the cache.}
}
\description{
Elimination matrix (memoised)
}
