% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_math_elementary.R
\name{exp,dual-method}
\alias{exp,dual-method}
\title{Element-wise exponential of a dual object}
\usage{
\S4method{exp}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Element-wise exponential of a dual object
}
