% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_diff.R
\name{finite_difference}
\alias{finite_difference}
\title{Finite differencing}
\usage{
finite_difference(f, vary, fix = NULL, h = 1e-08)
}
\arguments{
\item{f}{A function of which the derivative is seeked.}

\item{vary}{A named list of variables; the variables to be varied.}

\item{fix}{A named list of variables; the variables to be fixed.}

\item{h}{The finite differencing parameter; the size of perturbation.}
}
\description{
Finite differencing
}
\examples{
f <- function(y, X, beta) { y - X \%*\% beta }
finite_difference(f,
  vary = list(beta = c(5,6)),
  fix = list(X = matrix(1:4, 2, 2), y = c(2,3))
)

g <- function(X, Y) { X \%*\% Y }
finite_difference(g, vary = list(X = randn(2, 2), Y = randn(2, 2)))

}
