% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_math_elementary.R
\name{gamma,dual-method}
\alias{gamma,dual-method}
\title{Element-wise gamma of a dual object}
\usage{
\S4method{gamma}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Element-wise gamma of a dual object
}
