% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dual_def.R
\name{get_deriv}
\alias{get_deriv}
\title{Extract derivative}
\usage{
get_deriv(x, wrt)
}
\arguments{
\item{x}{A dual number}

\item{wrt}{A character string; the parameter name.}
}
\description{
Extract derivative
}
