% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dual_conversion.R
\name{\%<-\%}
\alias{\%<-\%}
\title{Single-level destructuring}
\usage{
lhs \%<-\% value
}
\arguments{
\item{lhs}{An expression; variables to be assigned.}

\item{value}{A list; values to assign.}
}
\description{
Strict usage: list(x, y) %<-% list(a, b, ...).
LHS must start with 'list'; RHS must be a list;
RHS must have at least as many elements as LHS.
}
\note{
Multiple-level is avoided intentionally, i.e. nested destructuring
is not allowed.
}
\examples{
\dontrun{
list(x, y) \%<-\% list(1, 2)
list(x, y) \%<-\% list(1, 2, 3)
}
}
