% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_arithmetic.R
\docType{methods}
\name{\%*\%,dual,ANY-method}
\alias{\%*\%,dual,ANY-method}
\title{Matrix multiplication of 'dual'-class objects}
\usage{
\S4method{\%*\%}{dual,ANY}(x, y)
}
\arguments{
\item{x}{A "dual" object.}

\item{y}{"ANY" object.}
}
\description{
Matrix multiplication of 'dual'-class objects
}
