% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dual.R
\docType{methods}
\name{\%*\%,dual,dual-method}
\alias{\%*\%,dual,dual-method}
\title{Matrix multiplication of 'dual'-class objects}
\usage{
\S4method{\%*\%}{dual,dual}(x, y)
}
\arguments{
\item{x}{A "dual" object.}

\item{y}{A "dual" object.}
}
\description{
Matrix multiplication of 'dual'-class objects
}
