% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_matrix_list.R
\docType{methods}
\name{\%*\%,matrix_list,matrix_list-method}
\alias{\%*\%,matrix_list,matrix_list-method}
\title{Matrix multiplication of 'matrix_list'-class objects}
\usage{
\S4method{\%*\%}{matrix_list,matrix_list}(x, y)
}
\arguments{
\item{x}{A "matrix_list" object.}

\item{y}{A "matrix_list" object.}
}
\description{
Matrix multiplication of 'matrix_list'-class objects
}
