% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_arithmetic.R
\docType{methods}
\name{\%x\%,ANY,dual-method}
\alias{\%x\%,ANY,dual-method}
\title{Kronecker multiplication of 'dual'-class objects}
\usage{
\S4method{\%x\%}{ANY,dual}(X, Y)
}
\arguments{
\item{X}{"ANY" object.}

\item{Y}{A "dual" object.}
}
\description{
Kronecker multiplication of 'dual'-class objects
}
