% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_matrix_list.R
\docType{methods}
\name{\%x\%,ANY,matrix_list-method}
\alias{\%x\%,ANY,matrix_list-method}
\title{Matrix kronecker multiplication of 'matrix_list'-class objects}
\usage{
\S4method{\%x\%}{ANY,matrix_list}(X, Y)
}
\arguments{
\item{X}{Any object other than "matrix_list" object.}

\item{Y}{A "matrix_list" object.}
}
\description{
Matrix kronecker multiplication of 'matrix_list'-class objects
}
