% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_arithmetic.R
\docType{methods}
\name{\%x\%,dual,ANY-method}
\alias{\%x\%,dual,ANY-method}
\title{Kronecker multiplication of 'dual'-class objects}
\usage{
\S4method{\%x\%}{dual,ANY}(X, Y)
}
\arguments{
\item{X}{A "dual" object.}

\item{Y}{"ANY" object.}
}
\description{
Kronecker multiplication of 'dual'-class objects
}
