% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_matrix_list.R
\docType{methods}
\name{\%x\%,matrix_list,ANY-method}
\alias{\%x\%,matrix_list,ANY-method}
\title{Matrix kronecker multiplication of 'matrix_list'-class objects}
\usage{
\S4method{\%x\%}{matrix_list,ANY}(X, Y)
}
\arguments{
\item{X}{A "matrix_list" object.}

\item{Y}{Any object other than "matrix_list" object.}
}
\description{
Matrix kronecker multiplication of 'matrix_list'-class objects
}
