% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_structure_subsetting.R
\name{Extract.dual}
\alias{Extract.dual}
\alias{[,dual,numeric,missing,ANY-method}
\alias{[,dual,missing,numeric,ANY-method}
\alias{[,dual,numeric,numeric,ANY-method}
\title{Extract parts of an object}
\usage{
Extract.dual(x, i, j, drop = FALSE)

\S4method{[}{dual,numeric,missing,ANY}(x, i, j, drop = FALSE)

\S4method{[}{dual,missing,numeric,ANY}(x, i, j, drop = FALSE)

\S4method{[}{dual,numeric,numeric,ANY}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{A "dual" object.}

\item{i}{integer; the row index.}

\item{j}{integer; the column index.}

\item{drop}{TRUE or FALSE; if TRUE, returns a vector when one
dimension of the matrix is 1.}
}
\description{
Extract parts of an object
}
