% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_derivative.R
\name{init_dx}
\alias{init_dx}
\title{Initialise the dual component, i.e. the storage matrices for derivatives}
\usage{
init_dx(dx_dims, dx_ind)
}
\arguments{
\item{dx_dims}{list of pairs; the list of matrix dimensions.}

\item{dx_ind}{An integer that indicates the location of the derivative
corresponding to x; input -1 if none.}
}
\description{
Initialise the dual component, i.e. the storage matrices for derivatives
}
\examples{
init_dx(list(c(4, 2), c(4, 5), c(4, 1)), -1)
init_dx(list(c(4, 2), c(4, 5), c(4, 1)), 1)
init_dx(list(c(4, 2), c(4, 5), c(4, 1)), 3)

}
