% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_sim_student_t.R
\name{inverse_transform}
\alias{inverse_transform}
\title{Simulation by inverse transform}
\usage{
inverse_transform(cdf)
}
\arguments{
\item{cdf}{A function; the distribution function of a random variable.}
}
\description{
Simulation by inverse transform
}
