% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_arithmetic_1_cross_class.R
\name{kronecker,ANY,dual-method}
\alias{kronecker,ANY,dual-method}
\title{Kronecker multiplication of 'dual'-class objects (ANY-dual)}
\usage{
\S4method{kronecker}{ANY,dual}(X, Y)
}
\arguments{
\item{X}{"ANY" object.}

\item{Y}{A "dual" object.}
}
\description{
Kronecker multiplication of 'dual'-class objects (ANY-dual)
}
