% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_arithmetic_2_same_class.R
\name{kronecker,dual,dual-method}
\alias{kronecker,dual,dual-method}
\title{Kronecker product of 'dual'-class objects (dual-dual)}
\usage{
\S4method{kronecker}{dual,dual}(X, Y)
}
\arguments{
\item{X}{A "dual" object.}

\item{Y}{A "dual" object.}
}
\description{
Kronecker product of 'dual'-class objects (dual-dual)
}
