% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_math_elementary.R
\name{log,dual-method}
\alias{log,dual-method}
\title{Element-wise logarithm of a dual object}
\usage{
\S4method{log}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Element-wise logarithm of a dual object
}
