% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_matrix_list_def.R
\docType{class}
\name{matrix-list-class}
\alias{matrix-list-class}
\title{S4 class "matrix_list"}
\description{
S4 class "matrix_list"
}
\section{Slots}{

\describe{
\item{\code{matrices}}{A list of matrices. Each matrix must be of class 'matrix' or
an object from the package "Matrix.}
}}

\examples{
library(ADtools)
a <- new("matrix_list", matrices = purrr::map(1:4, ~randn(2,2)))
}
