% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_structure_reshape.R
\docType{methods}
\name{matrix.dual}
\alias{matrix.dual}
\alias{matrix,dual-method}
\title{Create a matrix from a given set of values}
\usage{
\method{matrix}{dual}(data, nrow, ncol = 1, byrow = F,
  dimnames = NULL)

\S4method{matrix}{dual}(data = NA, nrow = 1, ncol = 1, byrow = F,
  dimnames = NULL)
}
\description{
Create a matrix from a given set of values

Coerce the first component of the dual object into a matrix.
}
