% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_matrix_ops.R
\name{matrix_determinant}
\alias{matrix_determinant}
\alias{det}
\alias{det.default}
\title{Determinant of a matrix}
\usage{
det(x, ...)

det.default(x, ...)
}
\arguments{
\item{x}{numeric matrix: logical matrices are coerced to numeric.}

\item{...}{Optional arguments.  At present none are used.  Previous
    versions of \code{det} allowed an optional \code{method} argument.
    This argument will be ignored but will not produce an error.}
}
\value{
For \code{det}, the determinant of \code{x}.  For \code{determinant}, a
  list with components
  \item{modulus}{a numeric value.  The modulus (absolute value) of the
    determinant if \code{logarithm} is \code{FALSE}; otherwise the
    logarithm of the modulus.}
  \item{sign}{integer; either \eqn{+1} or \eqn{-1} according to whether
    the determinant is positive or negative.}
}
\description{
\code{det} calculates the determinant of a matrix.  \code{determinant}
  is a generic function that returns separately the modulus of the determinant,
  optionally on the logarithm scale, and the sign of the determinant.
}
\details{
The \code{determinant} function uses an LU decomposition and the
  \code{det} function is simply a wrapper around a call to
  \code{determinant}.

  Often, computing the determinant is \emph{not} what you should be doing
  to solve a given problem.
}
\examples{
(x <- matrix(1:4, ncol = 2))
unlist(determinant(x))
det(x)

det(print(cbind(1, 1:3, c(2,0,1))))
}
