% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_matrix_list.R
\docType{methods}
\name{-,matrix_list,missing-method}
\alias{-,matrix_list,missing-method}
\title{Negation of 'matrix_list'-class objects}
\usage{
\S4method{-}{matrix_list,missing}(e1, e2)
}
\arguments{
\item{e1}{A "matrix_list" object.}

\item{e2}{Empty}
}
\description{
Negation of 'matrix_list'-class objects
}
