% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_matrix_chain_multiplication.R
\name{matrix_prod}
\alias{matrix_prod}
\title{Interface for optimal matrix chain multiplication}
\usage{
matrix_prod(..., method = "optimal")
}
\arguments{
\item{...}{Numeric matrices.}

\item{method}{"optimal" order or "natural" left-to-right order.}
}
\description{
Interface for optimal matrix chain multiplication
}
