% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_matrix_chain_multiplication.R
\name{mcm_optimal_order}
\alias{mcm_optimal_order}
\title{Find the optimal order of multiplying a matrix chain}
\usage{
mcm_optimal_order(x)
}
\arguments{
\item{x}{A numeric vector of matrix dimensions}
}
\description{
Find the optimal order of multiplying a matrix chain
}
