% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_math_sum.R
\name{mean,dual-method}
\alias{mean,dual-method}
\title{Mean of vector or matrix elements}
\usage{
\S4method{mean}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Mean of vector or matrix elements
}
