% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_special_matrix.R
\name{one_matrix}
\alias{one_matrix}
\title{Matrix of ones}
\usage{
one_matrix(nr, nc)
}
\arguments{
\item{nr}{integer; row dimension.}

\item{nc}{integer; column dimension.}
}
\description{
Matrix of ones
}
