% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_matrix_chain_multiplication.R
\name{optim_prod}
\alias{optim_prod}
\title{Executing the matrix multiplication given the optimal order}
\usage{
optim_prod(matrix_ls)
}
\arguments{
\item{matrix_ls}{A list of matrices to be multiplied}
}
\description{
Executing the matrix multiplication given the optimal order
}
