% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dual.R
\docType{methods}
\name{+,dual,dual-method}
\alias{+,dual,dual-method}
\title{(Scalar) Addition of 'dual'-class objects}
\usage{
\S4method{+}{dual,dual}(e1, e2)
}
\arguments{
\item{e1}{A "dual" object.}

\item{e2}{A "dual" object.}
}
\description{
(Scalar) Addition of 'dual'-class objects
}
