% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_arithmetic_2_same_class.R
\name{^,dual,dual-method}
\alias{^,dual,dual-method}
\title{Powers of 'dual'-class objects (dual-dual)}
\usage{
\S4method{^}{dual,dual}(e1, e2)
}
\arguments{
\item{e1}{A "dual" object.}

\item{e2}{A "dual" object.}
}
\description{
Powers of 'dual'-class objects (dual-dual)
}
