% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randu}
\alias{randu}
\title{Generate a matrix randomly from a uniform distribution}
\usage{
randu(..., min = 0, max = 1)
}
\arguments{
\item{...}{Any number of dimensions.}

\item{min}{lower limit of the distribution. Must be finite.}

\item{max}{upper limit of the distribution. Must be finite.}
}
\description{
Generate a matrix randomly from a uniform distribution
}
