% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_structure_binding.R
\name{rbind2,ANY,dual-method}
\alias{rbind2,ANY,dual-method}
\title{Combine 'dual'-class objects by Rows (ANY-dual)}
\usage{
\S4method{rbind2}{ANY,dual}(x, y)
}
\arguments{
\item{x}{ANY object.}

\item{y}{A "dual" object.}
}
\description{
Combine 'dual'-class objects by Rows (ANY-dual)
}
