% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_binding.R
\docType{methods}
\name{rbind2,dual,dual-method}
\alias{rbind2,dual,dual-method}
\title{Combine 'dual'-class objects by Rows.}
\usage{
\S4method{rbind2}{dual,dual}(x, y)
}
\arguments{
\item{x}{A "dual" object.}

\item{y}{A "dual" object.}
}
\description{
Combine 'dual'-class objects by Rows.
}
