% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_simulation.R
\name{rchisq0,numeric,dual-method}
\alias{rchisq0,numeric,dual-method}
\title{Simulate Chi-square random variates}
\usage{
\S4method{rchisq0}{numeric,dual}(n, df, method = "inv_tf")
}
\arguments{
\item{n}{Positive integer; number of observations.}

\item{df}{Degrees of freedom (can be non-integer).}

\item{method}{base or inv_tf; base refers to the function in the
`stats` package while inv_tf refers to inverse transform.}
}
\description{
Simulate Chi-square random variates
}
