% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_simulation.R
\name{rchisq0}
\alias{rchisq0}
\title{Simulate Chi-square random variates}
\usage{
rchisq0(n, df, method = "inv_tf")
}
\arguments{
\item{n}{Positive integer; number of observations.}

\item{df}{Degrees of freedom (can be non-integer).}

\item{method}{base or inv_tf; base refers to `stats::rchisq` while
inv_tf refers to inverse transform.}
}
\description{
Simulate Chi-square random variates
}
\examples{
n <- 10
rchisq0(n, df = 3)

}
