% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_simulation.R
\name{rexp0,numeric,dual-method}
\alias{rexp0,numeric,dual-method}
\title{Simulate exponential random variates}
\usage{
\S4method{rexp0}{numeric,dual}(n, rate = 1)
}
\arguments{
\item{n}{Positive integer; the number of samples.}

\item{rate}{A dual number; the rate of the exponential distribution.}
}
\description{
Simulate exponential random variates
}
