% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_simulation.R
\name{rgamma0}
\alias{rgamma0}
\title{Simulate gamma random variates}
\usage{
rgamma0(n, shape, scale)
}
\arguments{
\item{n}{Positive integer; the number of samples.}

\item{shape}{Positive number; the shape of the gamma distribution.}

\item{scale}{Positive number; the scale of the gamma distribution.}
}
\description{
Simulate gamma random variates
}
