% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_sim_normal.R
\docType{methods}
\name{rmvnorm0,numeric,dual,dual-method}
\alias{rmvnorm0,numeric,dual,dual-method}
\title{Simulate multivariate normal random variates}
\usage{
\S4method{rmvnorm0}{numeric,dual,dual}(n, mean, sigma)
}
\arguments{
\item{n}{Positive integer; the number of samples.}

\item{mean}{A dual number; the mean of the normal distribution.}

\item{sigma}{A dual number; the standard deviation of the normal distribution.}
}
\description{
Simulate multivariate normal random variates
}
