% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_sim_normal.R
\name{rmvnorm0_dual}
\alias{rmvnorm0_dual}
\alias{rmvnorm0,numeric,dual,dual-method}
\alias{rmvnorm0,numeric,dual,ANY-method}
\alias{rmvnorm0,numeric,ANY,dual-method}
\title{Simulate multivariate normal random variates}
\usage{
rmvnorm0_dual(n, mean, sigma)

\S4method{rmvnorm0}{numeric,dual,dual}(n, mean, sigma)

\S4method{rmvnorm0}{numeric,dual,ANY}(n, mean, sigma)

\S4method{rmvnorm0}{numeric,ANY,dual}(n, mean, sigma)
}
\arguments{
\item{n}{Positive integer; the number of samples.}

\item{mean}{A numeric vector or a dual number; the mean of the normal distribution.}

\item{sigma}{A numeric matrix or a dual number; the standard deviation of the normal distribution.}
}
\description{
Simulate multivariate normal random variates
}
