% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_sim_student_t.R
\name{rmvt0}
\alias{rmvt0}
\title{Simulate random variates from the multivariate t distribution}
\usage{
rmvt0(n, sigma, df, delta = 0)
}
\arguments{
\item{n}{positive integer; number of observations.}

\item{sigma}{scale matrix.}

\item{df}{degree of freedom.}

\item{delta}{non-centrality parameters.}
}
\description{
Simulate random variates from the multivariate t distribution
}
\examples{
n <- 10
d <- 3
rmvt0(n, sigma = crossprod(randn(d, d)), df = 2)

}
