% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_sim_normal.R
\docType{methods}
\name{rnorm0,numeric,dual,dual-method}
\alias{rnorm0,numeric,dual,dual-method}
\title{Simulate univariate normal random variates}
\usage{
\S4method{rnorm0}{numeric,dual,dual}(n, mean = 0, sd = 1)
}
\arguments{
\item{n}{Positive integer; the number of samples.}

\item{mean}{A dual number; the mean of the normal distribution.}

\item{sd}{A dual number; the standard deviation of the normal distribution.}
}
\description{
Simulate univariate normal random variates
}
