% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_generic.R
\name{round,dual,integer-method}
\alias{round,dual,integer-method}
\title{Rounding of Numbers}
\usage{
\S4method{round}{dual,integer}(x, digits = 0)
}
\arguments{
\item{x}{A "dual" object.}

\item{digits}{integer indicating the number of decimal places.}
}
\description{
Rounding of Numbers
}
