% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_generic.R
\name{round.dual}
\alias{round.dual}
\title{Rounding of Numbers}
\usage{
\method{round}{dual}(x, digits = 0)
}
\arguments{
\item{x}{A "dual" object.}

\item{digits}{integer indicating the number of decimal places.}
}
\description{
Rounding of Numbers
}
