% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_sim_student_t.R
\name{rt0,numeric,dual-method}
\alias{rt0,numeric,dual-method}
\title{Simulate random variates from the student-t distribution}
\usage{
\S4method{rt0}{numeric,dual}(n, df)
}
\arguments{
\item{n}{positive integer; the number of samples.}

\item{df}{A dual number; the degree of freedom.}
}
\description{
Simulate random variates from the student-t distribution
}
