% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_sim_student_t.R
\name{rt0}
\alias{rt0}
\title{Simulate random variates from the student-t distribution}
\usage{
rt0(n, df)
}
\arguments{
\item{n}{positive integer; the number of samples.}

\item{df}{degrees of freedom (> 0, maybe non-integer). df = Inf is allowed.}
}
\description{
Simulate random variates from the student-t distribution
}
\examples{
n <- 10
rt0(10, df = 3)

}
