% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_math_elementary.R
\name{sin,dual-method}
\alias{sin,dual-method}
\title{Element-wise sine of a dual object}
\usage{
\S4method{sin}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Element-wise sine of a dual object
}
