% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dual.R
\docType{methods}
\name{solve,dual,ANY-method}
\alias{solve,dual,ANY-method}
\title{Inverse of 'dual'-class objects}
\usage{
\S4method{solve}{dual,ANY}(a)
}
\arguments{
\item{a}{A "dual" object.}
}
\description{
Inverse of 'dual'-class objects
}
