% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_matrix_ops.R
\docType{methods}
\name{solve.dual}
\alias{solve.dual}
\alias{solve,dual,missing-method}
\alias{solve,dual,dual-method}
\title{Inverse of 'dual'-class objects}
\usage{
\method{solve}{dual}(a, b, ...)

\S4method{solve}{dual,missing}(a, b, ...)

\S4method{solve}{dual,dual}(a, b, ...)
}
\arguments{
\item{a}{A "dual" object.}

\item{b}{A "dual" object.}

\item{...}{Other arguments passed to 'base::solve'. See '?solve' for detail.}
}
\description{
Inverse of 'dual'-class objects
}
