% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_math_elementary.R
\name{sqrt,dual-method}
\alias{sqrt,dual-method}
\title{Element-wise square-root of a dual object}
\usage{
\S4method{sqrt}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Element-wise square-root of a dual object
}
