% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dual.R
\docType{methods}
\name{t,dual-method}
\alias{t,dual-method}
\title{Transpose of 'dual'-class objects}
\usage{
\S4method{t}{dual}(x)
}
\arguments{
\item{a}{A "dual" object.}
}
\description{
Transpose of 'dual'-class objects
}
