% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_structure_subsetting.R
\name{tail.dual}
\alias{tail.dual}
\alias{tail,dual-method}
\title{Return the First or Last Part of an Object}
\usage{
tail.dual(x, n = 6)

\S4method{tail}{dual}(x, n = 6)
}
\arguments{
\item{x}{A "dual" object.}

\item{n}{A single integer.}
}
\description{
Return the First or Last Part of an Object
}
