% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_math_elementary.R
\name{tan,dual-method}
\alias{tan,dual-method}
\title{Element-wise tangent of a dual object}
\usage{
\S4method{tan}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Element-wise tangent of a dual object
}
