% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_matrix_ops.R
\docType{methods}
\name{tcrossprod,dual,ANY-method}
\alias{tcrossprod,dual,ANY-method}
\title{Crossproduct of 'dual'-class objects}
\usage{
\S4method{tcrossprod}{dual,ANY}(x, y)
}
\arguments{
\item{x}{A "dual" object.}

\item{y}{Numeric matrix.}
}
\description{
Crossproduct of 'dual'-class objects
}
