% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dual.R
\docType{methods}
\name{tcrossprod,dual,dual-method}
\alias{tcrossprod,dual,dual-method}
\title{Crossproduct of 'dual'-class objects}
\usage{
\S4method{tcrossprod}{dual,dual}(x, y)
}
\arguments{
\item{a}{A "dual" object.}
}
\description{
Crossproduct of 'dual'-class objects
}
