% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_diff.R
\name{tidy_dx}
\alias{tidy_dx}
\title{Add rownames and colnames to the dual component}
\usage{
tidy_dx(x_dual, vary)
}
\arguments{
\item{x_dual}{A dual number}

\item{vary}{A named list of parameters}
}
\description{
Add rownames and colnames to the dual component
}
