% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_arithmetic.R
\docType{methods}
\name{*,ANY,dual-method}
\alias{*,ANY,dual-method}
\title{(Element-wise) Multiplication of 'dual'-class objects}
\usage{
\S4method{*}{ANY,dual}(e1, e2)
}
\arguments{
\item{e1}{"ANY" object.}

\item{e2}{A "dual" object.}
}
\description{
(Element-wise) Multiplication of 'dual'-class objects
}
