% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_arithmetic.R
\docType{methods}
\name{*,dual,ANY-method}
\alias{*,dual,ANY-method}
\title{(Element-wise) Multiplication of 'dual'-class objects}
\usage{
\S4method{*}{dual,ANY}(e1, e2)
}
\arguments{
\item{e1}{A "dual" object.}

\item{e2}{"ANY" object.}
}
\description{
(Element-wise) Multiplication of 'dual'-class objects
}
