% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_structure_reshape.R
\docType{methods}
\name{upper_tri_matrix}
\alias{upper_tri_matrix}
\alias{upper_tri_matrix,dual-method}
\title{Construct a lower triangular matrix from a vector}
\usage{
upper_tri_matrix(data, nrow = 1, ncol = 1, diag = F)

\S4method{upper_tri_matrix}{dual}(data, nrow = 1, ncol = 1, diag = F)
}
\arguments{
\item{data}{A numeric vector.}

\item{nrow}{A positive integer; the desired number of rows.}

\item{ncol}{A positive integer; the desired number of rows.}

\item{diag}{A logical; should the diagonal be included ?}
}
\description{
Construct a lower triangular matrix from a vector
}
