% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_diff.R
\name{vec_finite_diff}
\alias{vec_finite_diff}
\title{Finite differencing (Core)}
\usage{
vec_finite_diff(f, x, h = 1e-08)
}
\arguments{
\item{f}{A function of which the derivative is seeked.}

\item{x}{The point at which the derivative is seeked.}

\item{h}{The finite differencing parameter; the size of perturbation.}
}
\description{
Finite differencing (Core)
}
\examples{
\dontrun{
f <- function(x) { x[1] + 2* x[2] }
vec_finite_diff(f, c(2,3))
}
}
\keyword{internal}
