% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_simulation.R
\docType{methods}
\name{rWishart0,dual,dual-method}
\alias{rWishart0,dual,dual-method}
\alias{rWishart0,ANY,dual-method}
\alias{rWishart0,dual,ANY-method}
\title{Simulating from Wishart distribution using Bartlett decomposition}
\usage{
\S4method{rWishart0}{dual,dual}(v, M, method = "base")

\S4method{rWishart0}{ANY,dual}(v, M, method = "base")

\S4method{rWishart0}{dual,ANY}(v, M, method = "base")
}
\arguments{
\item{v}{A dual number or a scalar; degrees of freedom.}

\item{M}{A dual number or a matrix; the matrix parameter of the Wishart distribution.}

\item{method}{base or inv_tf; base refers to the function in the
`stats` package while inv_tf refers to inverse transform.}
}
\description{
Simulating from Wishart distribution using Bartlett decomposition
}
