% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_kronecker_product.R
\name{XBxA}
\alias{XBxA}
\title{Pre-multiplying a kronecker product}
\usage{
XBxA(X, B, A)
}
\arguments{
\item{X}{numeric matrix.}

\item{B}{numeric matrix.}

\item{A}{numeric matrix.}
}
\description{
Compute X (B \%x\% A)
}
