% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_special_matrix.R
\name{zero_matrix}
\alias{zero_matrix}
\title{Matrix of zeroes}
\usage{
zero_matrix(nr, nc)
}
\arguments{
\item{nr}{integer; row dimension.}

\item{nc}{integer; column dimension.}
}
\description{
Matrix of zeroes
}
\examples{
zero_matrix(1, 3)
zero_matrix(5, 1)
zero_matrix(2, 2)

}
