% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADtools.R
\docType{package}
\name{ADtools-package}
\alias{ADtools}
\alias{ADtools-package}
\title{ADtools: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
