% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_structure_reshape.R
\docType{methods}
\name{as.matrix,dual-method}
\alias{as.matrix,dual-method}
\title{Coerce the first component of the dual object into a matrix.}
\usage{
\S4method{as.matrix}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Coerce the first component of the dual object into a matrix.
}
