% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_special_matrix.R
\name{band_matrix0}
\alias{band_matrix0}
\title{Band matrix (memoised)}
\usage{
band_matrix0(...)
}
\arguments{
\item{...}{`nr` integer; number of rows.

`nc` integer; number of columns.

`x` A scalar or a vector to be placed on the diagonal of the matrix.

Use empty argument to clear the cache.}
}
\description{
Band matrix (memoised)
}
