% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_structure_binding.R
\name{cbind2,dual,ANY-method}
\alias{cbind2,dual,ANY-method}
\title{Combine 'dual'-class objects by Columns (dual-ANY)}
\usage{
\S4method{cbind2}{dual,ANY}(x, y)
}
\arguments{
\item{x}{A "dual" object.}

\item{y}{ANY object.}
}
\description{
Combine 'dual'-class objects by Columns (dual-ANY)
}
