% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_matrix_ops.R
\name{chol0}
\alias{chol0}
\title{Cholesky decomposition}
\usage{
chol0(x)
}
\arguments{
\item{x}{A numeric matrix.}
}
\description{
Cholesky decomposition
}
\note{
This function returns a lower-triangular matrix.
}
