% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dual.R
\docType{methods}
\name{crossprod,dual,missing-method}
\alias{crossprod,dual,missing-method}
\title{Crossproduct of 'dual'-class objects}
\usage{
\S4method{crossprod}{dual,missing}(x, y)
}
\arguments{
\item{a}{A "dual" object.}
}
\description{
Crossproduct of 'dual'-class objects
}
