% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_structure_reshape.R
\name{diag.dual}
\alias{diag.dual}
\alias{diag,dual-method}
\title{Diagonal matrix}
\usage{
diag.dual(x)

\S4method{diag}{dual}(x)
}
\arguments{
\item{x}{A "dual" object.}
}
\description{
Diagonal matrix
}
\note{
A single column matrix is treated as a vector. If one wants to call
\code{diag} on a single column matrix \code{x}, one can call
\code{x[1, 1]} instead.
}
