% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_density.R
\name{dinvWishart0}
\alias{dinvWishart0}
\title{The density of the inverse Wishart distribution}
\usage{
dinvWishart0(X, v, M, log = FALSE)
}
\arguments{
\item{X}{numeric matrix.}

\item{v}{degrees of freedom (> 0, maybe non-integer).}

\item{M}{numeric matrix; the scale matrix.}

\item{log}{logical; if TRUE, returns the log value.}
}
\description{
The density of the inverse Wishart distribution
}
\examples{
d <- 3
mat0 <- crossprod(randn(d, d))
X <- solve(rWishart0(v = 10, M = solve(mat0)))
dinvWishart0(X, v = 10, M = mat0)

}
