% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_density.R
\name{dnorm0}
\alias{dnorm0}
\title{The density of the normal distribution}
\usage{
dnorm0(x, mean, sd, log = FALSE)
}
\arguments{
\item{x}{vector or matrix of quantiles. If x is a matrix, each row is taken to be a quantile.}

\item{mean}{numeric vector; the mean vector.}

\item{sd}{numeric vector; the sd vector.}

\item{log}{logical; if TRUE, returns the log value.}
}
\description{
The density of the normal distribution
}
\examples{
n <- 10
x <- rnorm0(n, mean = 0, sd = 1)
dnorm0(x, mean = 0, sd = 1)

}
