% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dual_def.R
\docType{class}
\name{dual-class}
\alias{dual-class}
\title{S4 class "dual"}
\description{
This class attaches a dual component to a number / an array.
}
\section{Slots}{

\describe{
\item{\code{x}}{Any "numeric", i.e. vector, matrix or array.}

\item{\code{dx}}{matrix; also accepts any matrix classes from the "Matrix" package.}
}}

\examples{
a <- new("dual", x = randn(2,2), dx = 0)
b <- new("dual", x = randn(2,2), dx = init_dx(list(c(4, 1), c(4, 2)), 2))
}
