% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dual_def.R
\docType{class}
\name{dual-class}
\alias{dual-class}
\title{S4 class "dual"}
\description{
This class attaches a dual component to a number / an array.
}
\section{Slots}{

\describe{
\item{\code{x}}{Numeric matrix.}

\item{\code{dx}}{matrix; also accepts any matrix classes from the "Matrix" package.}
}}

\note{
Users should not construct the object directly, instead, use the
constructor helper `dual` provided.
}
