% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dual_def.R
\docType{class}
\name{dual-class}
\alias{dual-class}
\title{S4 class "dual"}
\description{
This class attaches a dual component to a number / matrix.
}
\section{Slots}{

\describe{
\item{\code{x}}{scalar, vector or matrix; also accepts any matrix classes from the "Matrix" package.}

\item{\code{dx}}{matrix; also accepts any matrix classes from the "Matrix" package.}

\item{\code{param}}{a named list, containing the column indices each variable occupies.}
}}

\note{
Users should not construct the object directly, instead, use the
constructor helper \link{dual} provided.
}
