% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_diff.R
\name{finite_diff}
\alias{finite_diff}
\title{Finite differencing}
\usage{
finite_diff(f, x, h = 1e-08)
}
\arguments{
\item{f}{A function of which the derivative is seeked.}

\item{x}{The point at which the derivative is required.}

\item{h}{The finite differencing parameter; the size of perturbation.}
}
\description{
Finite differencing
}
