% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_simulation.R
\docType{methods}
\name{rgamma0,numeric,dual,dual_or_numeric-method}
\alias{rgamma0,numeric,dual,dual_or_numeric-method}
\alias{rgamma0,numeric,numeric,dual-method}
\title{Simulate gamma random variates}
\usage{
\S4method{rgamma0}{numeric,dual,dual_or_numeric}(n, shape, scale)

\S4method{rgamma0}{numeric,numeric,dual}(n, shape, scale)
}
\arguments{
\item{n}{Positive integer; the number of samples.}

\item{shape}{A dual number or a scalar; the shape of the gamma distribution.}

\item{scale}{A dual number or a scalar; the scale of the gamma distribution.}
}
\description{
Simulate gamma random variates

Simulate gamma random variates
}
\note{
At least on of `shape` and `scale` should be a dual number.
Otherwise, `stats::rgamma` is called instead.
}
