% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%<-\%}
\alias{\%<-\%}
\title{Single-level destructuring}
\usage{
lhs \%<-\% rhs
}
\description{
Strict usage: c(x, y) %<-% list(a, b, ...).
LHS must start with 'c'; RHS must be a list;
RHS must have at least as many elements as LHS.
}
\note{
Multiple-level is avoided intentionally, i.e. nested destructuring
is not allowed.
}
\examples{
c(x, y) \%<-\% list(1, 2)
c(x, y) \%<-\% list(1, 2, 3)
}
