% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_arithmetic_1_cross_class.R
\name{\%*\%,ANY,dual-method}
\alias{\%*\%,ANY,dual-method}
\title{Matrix multiplication of 'dual'-class objects (ANY-dual)}
\usage{
\S4method{\%*\%}{ANY,dual}(x, y)
}
\arguments{
\item{x}{"ANY" object.}

\item{y}{A "dual" object.}
}
\description{
Matrix multiplication of 'dual'-class objects (ANY-dual)
}
