% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dual_construction.R
\name{init_dx}
\alias{init_dx}
\title{Initialise the dual component, i.e. the storage matrices for derivatives}
\usage{
init_dx(num_dim, denom_dim, num_ind)
}
\arguments{
\item{num_dim}{A number; the length of the numerator of a derivative.}

\item{denom_dim}{Numeric vector; the length of the denominators of a derivative.}

\item{num_ind}{An integer that indicates the location of the derivative
corresponding to x; input -1 if none.}
}
\description{
Initialise the dual component, i.e. the storage matrices for derivatives
}
\examples{
\dontrun{
init_dx(4, c(2, 5, 1), -1)
init_dx(4, c(2, 5, 1), 1)
init_dx(4, c(2, 5, 1), 2)
init_dx(4, c(2, 5, 1), 3)
}
}
\keyword{interal}
