% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_arithmetic_1_cross_class.R
\name{kronecker,dual,ANY-method}
\alias{kronecker,dual,ANY-method}
\title{Kronecker multiplication of 'dual'-class objects (dual-ANY)}
\usage{
\S4method{kronecker}{dual,ANY}(X, Y)
}
\arguments{
\item{X}{A "dual" object.}

\item{Y}{"ANY" object.}
}
\description{
Kronecker multiplication of 'dual'-class objects (dual-ANY)
}
