% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_structure_reshape.R
\name{matrix.dual}
\alias{matrix.dual}
\alias{matrix,dual-method}
\title{Coerce the first component of the dual object into a matrix.}
\usage{
\method{matrix}{dual}(data, nrow, ncol = 1, byrow = FALSE, dimnames = NULL)

\S4method{matrix}{dual}(data = NA, nrow = 1, ncol = 1, byrow = FALSE, dimnames = NULL)
}
\arguments{
\item{data}{A "dual" object.}

\item{nrow}{a positive integer; the desired number of rows.}

\item{ncol}{a positive integer; the desired number of rows.}

\item{byrow}{TRUE or FALSE; whether to fill the matrix by rows.}

\item{dimnames}{A dimnames attribute for the matrix: NULL or
a list of length 2 giving the row and column names respectively.
An empty list is treated as NULL, and a list of length one as
row names. The list can be named, and the list names will be
used as names for the dimensions.}
}
\description{
Coerce the first component of the dual object into a matrix.
}
