% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_matrix_ops.R
\name{matrix_determinant}
\alias{matrix_determinant}
\alias{det}
\alias{det.default}
\title{Determinant of a matrix}
\usage{
det(x, ...)

det.default(x, ...)
}
\arguments{
\item{x}{numeric matrix: logical matrices are coerced to numeric.}

\item{...}{Optional arguments.  At present none are used.  Previous
    versions of \code{det} allowed an optional \code{method} argument.
    This argument will be ignored but will not produce an error.}
}
\description{
Determinant of a matrix
}
