% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{one_matrix0}
\alias{one_matrix0}
\title{Matrix of ones}
\usage{
one_matrix0(...)
}
\arguments{
\item{nr}{integer; number of rows.}

\item{nc}{integer; number of columns.}

\item{x}{A scalar or a vector to be placed on the diagonal of the matrix.}
}
\description{
Matrix of ones
}
