% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{one_matrix0}
\alias{one_matrix0}
\title{Matrix of ones}
\usage{
one_matrix0(...)
}
\arguments{
\item{...}{`nr` integer; number of rows.

`nc` integer; number of columns.}
}
\description{
Matrix of ones
}
