% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_sim_normal.R
\name{rmvnorm0}
\alias{rmvnorm0}
\title{Simulate multivariate normal random variates}
\usage{
rmvnorm0(n, mean, sigma)
}
\arguments{
\item{n}{Positive integer; the number of samples.}

\item{mean}{mean vector of the normal distribution.}

\item{sigma}{covariance matrix of the normal distribution.}
}
\value{
A numeric matrix, where every column is a sample.
}
\description{
Simulate multivariate normal random variates
}
