% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rv_sim_normal.R
\name{rnorm0}
\alias{rnorm0}
\alias{rnorm0,numeric,dual,dual-method}
\alias{rnorm0,numeric,ANY,dual-method}
\alias{rnorm0,numeric,dual,ANY-method}
\title{Simulate univariate normal random variates}
\usage{
rnorm0(n, mean = 0, sd = 1)

\S4method{rnorm0}{numeric,dual,dual}(n, mean = 0, sd = 1)

\S4method{rnorm0}{numeric,ANY,dual}(n, mean = 0, sd = 1)

\S4method{rnorm0}{numeric,dual,ANY}(n, mean = 0, sd = 1)
}
\arguments{
\item{n}{Positive integer; the number of samples.}

\item{mean}{A dual number or a numeric vector; the mean of the normal distribution.}

\item{sd}{A dual number or a numeric vector; the standard deviation of the normal distribution.}
}
\description{
Simulate univariate normal random variates
}
