% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_generic.R
\name{round.dual}
\alias{round.dual}
\title{Rounding of Numbers}
\usage{
\method{round}{dual}(x, digits = 0)
}
\arguments{
\item{x}{A "dual" object.}

\item{digits}{An integer indicating the number of decimal places.}
}
\description{
Rounding of Numbers
}
\note{
The function 'round' does not have a derivative over the real line.
The derivative will be kept unchanged. The reason of not dropping it is
that sometimes one may need to round a matrix to correct floating-point
errors. This is often used when an apparent symmetric matrix does not pass
the symmetry test.
}
