% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_matrix_ops.R
\name{solve,dual,missing-method}
\alias{solve,dual,missing-method}
\alias{solve,dual,dual-method}
\alias{solve,ANY,dual-method}
\alias{solve,dual,ANY-method}
\title{Inverse of 'dual'-class objects}
\usage{
\S4method{solve}{dual,missing}(a, b, ...)

\S4method{solve}{dual,dual}(a, b, ...)

\S4method{solve}{ANY,dual}(a, b, ...)

\S4method{solve}{dual,ANY}(a, b, ...)
}
\arguments{
\item{a}{A numeric square matrix or a corresponding "dual" object.}

\item{b}{A numeric vector or matrix or a corresponding "dual" object.}

\item{...}{Other arguments passed to 'base::solve'. See '?solve' for detail.}
}
\description{
Inverse of 'dual'-class objects
}
\note{
At least one of `a` and `b` must be a dual object, and the other one 
may be any appropriate numeric matrix.
}
