% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_structure_reshape.R
\name{upper_tri_matrix}
\alias{upper_tri_matrix}
\alias{upper_tri_matrix,dual-method}
\title{Construct an upper triangular matrix from a vector}
\usage{
upper_tri_matrix(data, nrow = 1, ncol = 1, diag = FALSE)

\S4method{upper_tri_matrix}{dual}(data, nrow = 1, ncol = 1, diag = FALSE)
}
\arguments{
\item{data}{A numeric vector.}

\item{nrow}{A positive integer; the desired number of rows.}

\item{ncol}{A positive integer; the desired number of rows.}

\item{diag}{A logical; should the diagonal be included ?}
}
\description{
Construct an upper triangular matrix from a vector
}
\examples{
upper_tri_matrix(1:3, 3, 3)
upper_tri_matrix(1:6, 3, 3, diag = TRUE)

}
