% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_special_matrix.R
\name{zero_matrix0}
\alias{zero_matrix0}
\title{Matrix of zeroes (Memoised)}
\usage{
zero_matrix0(...)
}
\arguments{
\item{...}{`nr` integer; number of rows.

`nc` integer; number of columns.}
}
\description{
Matrix of zeroes (Memoised)
}
